First Truly Working Interference Code
interference.py is the main code in this commit. It requires a folder crms in the Data directory to take control amplitudes. This is done by calling get_data.py and that file pulling the crms data. new_mainOverlay.png is used by interference.py to generate the walls of the great hall that are used to calculate the normal line of the walls for the sound bounce. Must also create srms_bounce and srms_direct folders in the data directory before running.